requirements.txt
tensorflow
gym==0.10.11
tqdm
numpy
matplotlib
PILLOW
pyglet==1.3.2
pyvirtualdisplay
tf-agents

